refactoring: graphQL yoga clean code
* 사용하지 않는 jwt.js 파일 삭제
* pollGuest.resolver.js pollGuest.resolver.js의 공통 함수를 resolveHelper.js로 이동
* 실행 흐름 구분을 위한 줄바꿈 추가
* 불필요한 할당 제거
* 인증 미들웨어를 파일로 분리
* context 함수의 로직개선
* fix lint
